+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Rely on Proton to declare common annotation processor options.
+
 ### [v0.47](https://github.com/realityforge/router-fu/tree/v0.47) (2026-06-16) · [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.46...v0.47)
 
 Changes in this release: